make mint description customizable and move seatMint to core/
